Use proper name for $rid parameter. (#5898)
It's been brought to our attention that we use incorrect parameter name. It should be `$rid`. 
**Note: this change only affect tests. The actual topic translation functionality works fine.**
I propose this test just to avoid confusion in the future where someone can look at the test and get the wrong impression about parameter name.

This PR does not require additional test coverage.

## Azure IoT Edge PR checklist: