Use proper name for $rid parameter. (#5898) (#5909)
It's been brought to our attention that we use incorrect parameter name. It should be `$rid`. 
**Note: this change only affect tests. The actual topic translation functionality works fine.**
I propose this test just to avoid confusion in the future where someone can look at the test and get the wrong impression about parameter name.

This PR does not require additional test coverage.

## Azure IoT Edge PR checklist:

This checklist is used to make sure that common guidelines for a pull request are followed.

### General Guidelines and Best Practices
- [x] I have read the [contribution guidelines](https://github.com/azure/iotedge#contributing).
- [x] Title of the pull request is clear and informative.
- [x] Description of the pull request includes a concise summary of the enhancement or bug fix.