Port parallelization changes from images pipeline to addons-release (#5904)
Port the parallelization changes from the image builder pipeline to the addons-release pipeline. This fixes the addons-release pipeline, which is currently broken because it depends on build-api-proxy.yaml, which was changed from a template containing   'tasks' to one containing 'jobs' as part of the parallelization changes to the image builder pipeline.

The changes have been tested with a successful test run of the addons-release pipeline. I then tested the resulting build of the API Proxy image by using it in a nested edge deployment and verifying that it worked; the lower layer device was able to pull the diagnostics image via the API proxy running on the top layer device.

## Azure IoT Edge PR checklist:

This checklist is used to make sure that common guidelines for a pull request are followed.

### General Guidelines and Best Practices
- [x] I have read the [contribution guidelines](https://github.com/azure/iotedge#contributing).
- [x] Title of the pull request is clear and informative.
- [x] Description of the pull request includes a concise summary of the enhancement or bug fix.

### Testing Guidelines
- [x] Pull request includes test coverage for the included changes.
- Description of the pull request includes 
	- [x] concise summary of tests added/modified
	- [x] local testing done.  

### Draft PRs
- Open the PR in `Draft` mode if it is:
	- Work in progress or not intended to be merged.
	- Encountering multiple pipeline failures and working on fixes.

_Note: We use the kodiakhq bot to merge PRs once the necessary checks and approvals are in place. When it merges a PR, kodiakhq converts the PR title to the commit title, PR description to the commit description, and squashes all the commits in the PR to a single commit. The net effect is that entire PR becomes a single commit. Please follow the best practices mentioned [here](https://chris.beams.io/posts/git-commit/#:~:text=The%20seven%20rules%20of%20a%20great%20Git%20commit,what%20and%20why%20vs.%20how%20For%20example%3A%20) for the PR title and description_