Cherry-pick: Port parallelization changes from images pipeline to addons-release (#5923)
Port the parallelization changes from the image builder pipeline to the addons-release pipeline. This fixes the addons-release pipeline, which is currently broken because it depends on build-api-proxy.yaml, which was changed from a template containing   'tasks' to one containing 'jobs' as part of the parallelization changes to the image builder pipeline.
    
The changes have been tested with a successful test run of the addons-release pipeline. I then tested the resulting build of the API Proxy image by using it in a nested edge deployment and verifying that it worked; the lower layer device was able to pull the diagnostics image via the API proxy running on the top layer device.

## Azure IoT Edge PR checklist: