[E2E]: Enable aziot-edge install from Microsoft prod (#4881) (#5941)
Straight cherry-pick:  This change will enable aziot-edge install from Microsoft prod. This is useful for scenarios where we don't care about specific version of runtime, but are fine with the latest officially published.

The use case for 1.2 is to make the release process simpler, when it comes to base images updates. 

- [X] I have read the [contribution guidelines](https://github.com/azure/iotedge#contributing).