Build rocksdb and arm images in amd64 hosts (ubuntu 20.04 hosts) (#5947)
When this project was started, we used base images to run some commands in the native execution environment (for example: create a user in the container).  Now with cross-platform docker builds, we can use runtime emulation to do this work, and not maintain the base images.

We also had an inconsistent way to house the RocksDB library files.  This PR builds the libraries in parallel with the project executables, and collects the libraries with the other artifacts for the image build to use. I also updated the container images build to ubuntu 20.04 to remove all references to qemu-static container.

This PR reduces (but does not completely eliminate) dependency on docker hub for the project.

Testing for this PR was done through executing E2E tests with images built from this PR.