Rocksdb build for arm needs Ubuntu 18 (#5950)
It looks like if I use ubuntu 20.04 as a base for the RocksDB build,
it pulls in the incorrect version of glibc. The dotnet containers are 
based on bionic, so dropping down to 18.04 should
allow the libsrocksdb to be loaded.

Ran E2E test again, and checked for the exception I saw earlier. The storage is created cleanly.


## Azure IoT Edge PR checklist: