CI: Disable flaky SendTelemetry test (#5945)
This CI test is flaky, but is also redundant as the scenario is also covered in our E2E suite. 

The root cause of the flake is that when the module client is disconnected, sometimes the amqp session is no longer there so the SDK will throw an exception when closing the ModuleClient. We could refactor this test to avoid this, but it seems less error-prone to rely on just the existing E2E test to cover this scenario.

Test run with changes:
https://dev.azure.com/msazure/One/_build/results?buildId=50188304&view=results

***Please replace this line with your PR description and read PR checklist below***

## Azure IoT Edge PR checklist: