Disable chaining verification in EA and EH (#5953)
Disable chaining verification in EA and EH as there is known bug in the way the certs are processed after a deployment is set. Util a fix is made, disabling it for now. Also the chaining verification is done already in the Manifest signer client so it doesn't needed for now.