Update Microsoft.Azure.Devices.Client from 1.36.3 to 1.36.4 (#6115) (#6120)
Update Microsoft.Azure.Devices.Client version to 1.36.4 which includes fix for dotnetty bug where it created threads in foreground instead of background. The bug in dotnetty caused issue with certificate renewal in edgeHub because the process is not terminated when UpstreamProtocol is Mqtt.
Tested manually that certificate renewal works.

## Azure IoT Edge PR checklist:


## Azure IoT Edge PR checklist:

This checklist is used to make sure that common guidelines for a pull request are followed.

### General Guidelines and Best Practices
- [x] I have read the [contribution guidelines](https://github.com/azure/iotedge#contributing).
- [x] Title of the pull request is clear and informative.
- [x] Description of the pull request includes a concise summary of the enhancement or bug fix.

### Testing Guidelines
- [x] Pull request includes test coverage for the included changes.
- Description of the pull request includes 
	- [x] concise summary of tests added/modified
	- [x] local testing done.