Increase test timeout for Windows (#6123)
The TestGetModuleLogs test occasionally fails, not because it does something wrong but because it simply doesn't have time to finish all the stuff it's trying to do. This change increases the end-to-end test timeout to 8 minutes for Windows. It also increases the overall test time allotment to 120 minutes (default is 60). We really don't need to double the allotment, but arm32v7 will already take this long so it won't hurt to double (as opposed to picking some arbitrarily smaller value).

TestGetModuleLogs has only failed twice for this reason in the last couple of weeks, so it's hard to know whether this corrects the problem. But as long as this doesn't break anything, we'll make this change and see if flaky runs disappear over time.

## Azure IoT Edge PR checklist:

This checklist is used to make sure that common guidelines for a pull request are followed.

### General Guidelines and Best Practices
- [X ] I have read the [contribution guidelines](https://github.com/azure/iotedge#contributing).
- [X] Title of the pull request is clear and informative.
- [X] Description of the pull request includes a concise summary of the enhancement or bug fix.

### Testing Guidelines
- [X] Pull request includes test coverage for the included changes.
- Description of the pull request includes 
	- [X] concise summary of tests added/modified
	- [X] local testing done.