add dotnetty.common reference to 0.7.1 (#6122)
Reference dotnetty.common 0.7.1 to force using this version because it has a fix for bug introduced in 0.7.0 which client SDK uses. Bug causes edgeHub process to not be terminated for certificate renewal when UpstreamProtocol is Mqtt

Tested manually that edgeHub shutdown works when UpstreamProtocol is Mqtt and added e2e test for Mqtt.

## Azure IoT Edge PR checklist:

This checklist is used to make sure that common guidelines for a pull request are followed.

### General Guidelines and Best Practices
- [x] I have read the [contribution guidelines](https://github.com/azure/iotedge#contributing).
- [x] Title of the pull request is clear and informative.
- [x] Description of the pull request includes a concise summary of the enhancement or bug fix.

### Testing Guidelines
- [x] Pull request includes test coverage for the included changes.
- Description of the pull request includes 
	- [x] concise summary of tests added/modified
	- [x] local testing done.