Update scripts for removing keys and certificates on edge device (#6125)
The purpose of this PR is to change the scripts for removing the keys and certificates existing on the edge device.
If a user wants to run the current scripts using 'sudo', which is often the case, they would not work since the wildcard expansion is done by the shell, not by 'rm'.  Given that the shell does not have permission to list the directory '/var/lib/aziot/keyd/keys/', there is no expansion of the wildcard so it will not "list" any files and therefore 'rm' won't try to delete anything.

I've also added 'sudo' to all other commands in order to have unity throughout the file.


## Azure IoT Edge PR checklist: