[Tools] Check for shared library dependency in Platform validation tool (#6162)
[](https://user-images.githubusercontent.com/33068020/153299807-5dcb3058-ff2e-403d-9308-202e497701fb.png)[](https://user-images.githubusercontent.com/33068020/153298960-bc88d1ba-3d5a-4122-9afa-2a7150f4644d.png)he shared libraries that are part of OS distribution is checked by this tool.

The list of shared libraries that are needed for aziot-edge and aziot-identity-service are as follows

libcrypto.so.1.1
libssl.so.1.1
libdl.so.2
libpthread.so.0
libgcc_s.so.1
libc.so.6
ld-linux-x86-64.so.2
libm.so.6

Notes:

libaziot_keys.so is not part of OS but it's installed with aziot-identity-service. so skipped it in this check.
Specific version is checked as there is no minimum version research has been done with which IoT edge runtime works perfectly. Always using the latest version from the distribution and the tools needs to keep up to date with this.
Example of all checks passing
![image](https://user-images.githubusercontent.com/33068020/155640687-ac8bb542-b50d-4ca2-b4e9-1bdf6983476f.png)


Example scenario of missing libssl.so.1.1
![image](https://user-images.githubusercontent.com/33068020/155640702-12c74849-4297-4c5b-bbde-bc077fc69369.png)