Add Memory and CPU Usage Analysis Script for Platform Compatibility (#6163)
One of the Core Requirements for Platform Compatibility tool is to add a soft check letting users know about the memory and CPU Usage of IoT Edge Runtime. The goal of this script is to run it as part of a release process to get usage statistics that will be added to the platform compatibility tool with every release. The script uses a combination of top output and docker stats.

The output from this script is a file with the following snippet. 

```sh
aziot-edged-size=10.97
aziot-edged-peak-memory=15.82
aziot-edged-avg-memory=14.38
aziot-edged-peak-cpu=35.3
aziot-edged-avg-cpu=0.01
aziot-identityd-size=8.57
```
The Next Goal would be

1. Integrate this script in a bench setup for release.
2. Add Memory Check in Platform compatibility tool to use these statistics to perform a soft check
3. Automatically update platform compatibility tool with updated numbers with every release

## Azure IoT Edge PR checklist: