Revert usage of arsing/cross to upstream cross 0.1
The fix we needed was merged into upstream a long time ago.

Ref: https://github.com/cross-rs/cross/issues/52
Ref: https://github.com/cross-rs/cross/commit/249500e9a1f953d7eeca8a8b744e3aa1a10f64e6

The image-linux-rust job uses cross 0.1 and not 0.2. This is because
this job builds iotedge-proxy for x86_64-unknown-linux-musl, to make
an iotedge-proxy Docker image for Kubernetes. We do not have
a custom builder image for this target, so cross 0.2 would use upstream's image,
which does not have openssl. We *could* use cross 0.2 with a custom image,
like we do for the ARM targets (see Cross.toml), but for now this change just
sticks with cross 0.1. The openssl in 0.1's image is still ancient
(the image hasn't been updated in 4 years), but Dave said that can be tackled
later, as part of updating openssl across this branch in general.