Use 1ES-hosted agent pool for arm64 end-to-end tests (#6076)
This change moves our arm64 end-to-end tests to use preview 1ES-hosted agent pools instead of Raspberry Pis. The job runs a lot faster, and it reduces the physical hardware we have to maintain. The tradeoff is that, while arm64 agents are in preview, there are limitations we have to work around by adding ugly script directly into the job yaml. We should be able to remove it soon, when arm64 agents become generally available.

I ran the end-to-end test pipeline a few times and made sure it passed.

## Azure IoT Edge PR checklist:

This checklist is used to make sure that common guidelines for a pull request are followed.

### General Guidelines and Best Practices
- [X] I have read the [contribution guidelines](https://github.com/azure/iotedge#contributing).
- [X] Title of the pull request is clear and informative.
- [X] Description of the pull request includes a concise summary of the enhancement or bug fix.

### Testing Guidelines
- [X] Pull request includes test coverage for the included changes.
- Description of the pull request includes 
	- [X] concise summary of tests added/modified
	- [X] local testing done.