Fix openssl build on windows (#6280)
When building edgelet, the HSM components need a local installation of OpenSSL to link against. In our "iotedged Checkin" pipeline (and probably elsewhere) we build OpenSSL from source using vcpkg. A couple weeks ago, vcpkg updated their OpenSSL scripts to build 3.0 instead of 1.1.1. We see build errors in our pipelines when attempting to link against OpenSSL 3.0.

This change pins the vcpkg port of openssl to a specific commit (1.1.1n). The approach is recommended [here](https://vcpkg.io/en/docs/about/faq.html#how-do-i-use-different-versions-of-a-library-on-one-machine). This allows us to continue building edgelet the way we were before. We only need to do this in `release/1.1` as our architecture is quite different in 1.2.

## Azure IoT Edge PR checklist: