Mariner 2.0 Package Build for IoTEdge (#6255)
Checking if number of commits was causing dotnet checks to fail in main PR located here https://github.com/Azure/iotedge/pull/6224. This PR will replace the previous one, since this can pass the dotnet checks.

This PR will add a Mariner 2.0 build for Iotedge version 1.2.x. One Issue with this is that Mariner does not directly support rust 1.47.0, to fix this issue the rust toolchain from the build agent is included as an artifact in the spec file. This removes a dependency on the rust dependency for both Mariner 1.0 and Mariner 2.0, since all rust artifacts are now self contained in the spec files artifacts.

The changes also allow for the Mariner package build to select multiple different versions of the Mariner toolkit, and se the appropriate one for each package build. 

The resulting package was tested on a Mariner 2.0 environment and confirmed the functionality of IotEdge.

## Azure IoT Edge PR checklist: