Changes for obsagent testability (#6337)
Minor changes to azure-iot-mqtt and mqtt3 crates to:
1. Make some structs publicly accessible (for use outside the crate)
2. Make twin related structs cloneable

These changes are needed to ease unit testing of the observability agent. The obs agent currently depends on these iotedge crates for sending mqtt messages to IoT Hub. We're currently adding tests that mock out the functionality provided by the azure-iot-mqtt and mqtt3 crates. The changes in this PR will make that easier.

## Azure IoT Edge PR checklist: