Remove Debian 9 from main (#6332)
Debian 9 is EOL and we are planning to drop support for it in the next LTS. As part of this, we need to remove Deb 9 arm32 jobs from our build, test, and release pipelines (only in branch 'main').

This PR removes the build/release of Debian 9 artifacts and replaces Debian 9 artifacts with Debian 10, in our existing test pipelines.

**Testing notes:**

Build Images, End-to-End tests, Long Haul, Connectivity, and Edgelet packages pipelines passed.

- [X] I have read the [contribution guidelines](https://github.com/azure/iotedge#contributing).