Mariner ARM64 Build [Squashed] (#6356)
This PR adds a aarch64 build for Mariner 1.0 and 2.0 for iotedge 1.2. Mariner is not able to cross-compile so it requires a native aarch64 device to create the rpm artifacts. Due being unable to cross-compile we also need to build iot-identity-service in this pipelines since iot-identity-service does not support building on non-amd64 devices.

Same as this PR https://github.com/Azure/iotedge/pull/6325 , but has been squashed to prevent test failure.

## Azure IoT Edge PR checklist: