Test upgrade scenarios 1.1 (#6367)
Adds an ADO pipeline for testing various partial upgrade scenarios, all with a 1.1.x edge daemon installed (this is for simplicity so that the 1.1 .NET E2E test framework could be used). Other scenarios are covered by similar pipelines on the other release branches.

These changes were tested by running test pipelines on the PR branch and validating the final parameters in context.json for each scenario as well as spot checking the support bundles to confirm the correct versions of aziot-edged and aziot-identity-service are installed during the run.

TODO in follow-on/backlogged PRs:
1. Test full upgrade scenarios
2. Reduce code duplication in upgrade-scenarios.yaml

## Azure IoT Edge PR checklist: