[1.3] Lock cargo tarpaulin upstream dependences to a specific version (#6394)
cargo tarpaulin doesn't pin its upstream dependences to certain versions.

This runs the risk that a change/update to an upstream dependency might be pulled in and end up breaking the cargo tarpaulin compile. This isn't a problem right now, but with 1.3 being cut, this is likely to be a problem down the line.

This PR ensures that only locked upstream dependences are pulled in for the cargo tarpaulin compile.

- [X] I have read the [contribution guidelines](https://github.com/azure/iotedge#contributing).