Lock cargo tarpaulin upstream dependences to a specific version (#6392) (#6393)
cargo tarpaulin is pinned to the latest 0.18.x version in release/1.1

However, it doesn't pin its upstream dependences to certain versions. 

What's currently happening is that a change to[ libz-sys ](https://crates.io/crates/libz-sys/versions)5 days ago, has caused that version (1.1.18) to be pulled in and break the cargo tarpaulin compile.

This PR ensures that only locked upstream dependences are pulled in for the cargo tarpaulin compile.