Disable SCCA task (#6402)
The SCCA task can fail the builds if it detects Dockerfiles in our repo that reference external images. We're working on a solution with the SCCA team that will meet the needs of our open-source project. In the meantime, we're disabling this task on their recommendation.

I used the Build Images pipeline to confirm that the SCCA task passes when the `DisableDockerDetector` variable is set in the pipeline.

## Azure IoT Edge PR checklist: