Mitigate upload support bundle test issues (#6433)
Mitigates the failure likelihood for the UploadSupportBundle E2E test by reducing the amount of data being uploaded. This should reduce the likelihood of timeout on resource constrained agents such as 32-bit Raspberry Pis. 

Also, the response and connection timeouts for both the UploadSupportBundle and GetTaskStatus direct methods have been set to the max allowed value of 300 seconds. This aligns with the values on the 1.2 branch.

These changes build on top of #6424, which allowed the tests to get farther (i.e. edgelet now gets the supportbundle GET request that it was not getting previously in observed failure scenarios).

I've run the pipeline 3 times and it succeeded each time.

## Azure IoT Edge PR checklist: