Increase E2E test direct method timeouts to 300 sec (#6436)
Mitigates E2E test failures on resource constrained agents such as the 32-bit Raspberry Pis. 

The timeout values on other branches (i.e. release/1.2 and release/1.1) are already set to 300, so this change aligns the main branch with the release branches.

A test pipeline was run and it was confirmed that all direct method tests passed.

## Azure IoT Edge PR checklist: