[main] Upgrade to latest Rust version (#6440)
This is our monthly toolchain upgrade to the latest stable version. Since iotedge/edgelet is a binary package, we want to be tracking Rust compiler releases closely in the event a stdlib vulnerability is found. We do not pin to "stable", however, since that has caused pipeline breakage when some code patterns are made illegal (like in the upgrade from 1.47 to 1.48 [^0]) or, more often, clippy lints are added.

- Switch `.to_owned` to `.clone` where semantically appropriate (`clippy::implicit_clone`)
- Merge pattern matching statements (`clippy::unnested_or_patterns`)
- Stop immediately dropping `Drop` types (`clippy::let_underscore_drop`)
- Use semicolon for unit returns (`clippy::semicolon_if_nothing_returned`)
- Remove needless borrows (`clippy::needless_borrow`)
- Stop using `.to_string` in format arguments (`clippy::to_string_in_format_args`)
- Use `.unwrap_or_else` where function calls required (`clippy::or_fun_call`)

[^0]: https://github.com/rust-lang/rust/blob/master/RELEASES.md#compatibility-notes-11
  Namely, the point on `mem::uninitialized`.

## Azure IoT Edge PR checklist: