open ot proxy (#6439)
otproxy is always closed from the network. This prevent security patches from being applied.

With this PR, otproxy will be open by default and locked only when performing the tests. This way, most of the time OT proxy will be open to new security patches.


Lock
Checked that the VM is locked in portal

Unlock
Checked that the VM is unlocked in portal.

https://dev.azure.com/msazure/One/_build/results?buildId=56491150&view=logs&j=c27c4ff5-2c93-5622-6650-0f5d00fc53b7&t=84dfacbf-177f-5115-f3f0-0f5e9b223605