[main] Flatten additional properties of metrics (#6322)
The recent PR for adding arbitrary device product information (#6262) changed the structure of Microsoft.Azure.Devices.Edge.Agent.Core.SystemInfo, which in turn affected the metrics sent upstream.  While it is no longer feasible with the current regime of passing through extra properties to restore the old property names verbatim, we can at least flatten the additional properties to restore the general shape of the serialized information.  There is also a new test verifying that the flattening occurs as expected.

## Azure IoT Edge PR checklist: