Rename dockersource to gnarly (#6545)
We recently started using the dockersource tool to generate the `--build-context` argument that gets passed to `docker buildx build` in our Docker image builds pipeline. Last week, it was moved to a new repo and renamed. This change updates our references to the tool. Without this change, image builds will fail because these scripts refer to a tool that no longer exists by the name 'dockersource'.

I ran the Build Images pipeline against these changes to exercise the script changes.