Remove snitcher test module (#6521)
In v1.1, we use the snitcher module to email reports to the team during our long-running test pipelines (long haul and stress). We removed it from our main branch a while back because we invested in a replacement that logs to a dashboard. We aren't going to port the replacement to this branch, but that's okay because we can manually query the reporting endpoint instead of waiting for an email. Also, the module is old, we no longer build it, and the latest Docker images report vulnerabilities.

The stress test and base image update pipelines were also decommissioned long ago, so this change removes them rather than surgically extracting snitcher references from them.

## Azure IoT Edge PR checklist: