Socket Activation for Mariner Package Builds (#6526)
Originally Mariner was using the same method to add mgmt and work sockets as centos however this was not necessary since Mariner uses a newer version of systemd that allows for socket activation. This change has the mariner spec file incorporate .socket files and the .service file that requites both socket files.

## Azure IoT Edge PR checklist: