support privileged modules specified without CAP_ (#6182)
In a previous security fix, the allow_privileged flag can be specified to automatically remove privileged permissions from modules. If this is true, we add CAP_CHOWN and CAP_SETUID to the CapDrop host config field. 

For these permissions, docker automatically adds CAP_ to a permission if it is left off. With the old implementation, specifying that a module was allowed to have CHOWN would result in edgelet still adding CAP_CHOWN to the drop field. From docker's perspective, this is both adding and removing the same permission. 

I believe if a permission is in both CapAdd and CapDrop it will be dropped, but it is still better to avoid this scenario in the first place.