Fix packages pipeline CI trigger (#6617)
The edgelet packages pipeline runs every time a commit is merged to main. It should also run when commits are merged to release/1.4. This change updates the trigger spec to make that happen. Once this change is merged into main I'll also make sure it goes to release/1.4, since that's where the trigger will be evaluated.

## Azure IoT Edge PR checklist: